test: fixed minor issues with the new implementation

+++ dojo-starter/src/models/harvest_action.cairo
@@ -5,5 +5,5 @@ pub struct HarvestAction {
     pub entity: u32,
     pub start_ts: u64,
     pub end_ts: u64,
-    pub harvest_amount: u64,
+    pub harvest_amount: u128,
 }
